packages/latex2omml/pyproject.toml
[project]
name = "latex2omml"
version = "0.1.0"
description = "Convert LaTeX math to native Word equations (OMML) via python-docx"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Hosung You"}
]
keywords = ["latex", "omml", "word", "equations", "python-docx", "academic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "lxml>=4.0",
    "python-docx>=0.8.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/latex2omml"]
